src/main/webapp/app/entities/sensor-model/route/sensor-model-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ISensorModel, SensorModel } from '../sensor-model.model';
import { SensorModelService } from '../service/sensor-model.service';

import { SensorModelRoutingResolveService } from './sensor-model-routing-resolve.service';

describe('SensorModel routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: SensorModelRoutingResolveService;
  let service: SensorModelService;
  let resultSensorModel: ISensorModel | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(SensorModelRoutingResolveService);
    service = TestBed.inject(SensorModelService);
    resultSensorModel = undefined;
  });

  describe('resolve', () => {
    it('should return ISensorModel returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSensorModel = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSensorModel).toEqual({ id: 123 });
    });

    it('should return new ISensorModel if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSensorModel = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultSensorModel).toEqual(new SensorModel());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as SensorModel })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultSensorModel = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultSensorModel).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
